Move volume label creation into geo params
