+++ src/geocel/GeantGeoUtils.cc
@@ -10,7 +10,6 @@
 #include <iostream>
 #include <string>
 #include <string_view>
-#include <unordered_map>
 #include <unordered_set>
 #include <G4Element.hh>
 #include <G4FieldManager.hh>
@@ -42,8 +41,6 @@
 
 #include "GeoParamsInterface.hh"
 #include "g4/VisitVolumes.hh"
-
-#include "detail/MakeLabelVector.hh"
 
 // Check Geant4-reported and CMake-configured versions, mapping from
 // Geant4's base-10 XXYZ -> to Celeritas base-16 0xXXYYZZ
@@ -297,63 +294,6 @@ find_geant_volumes(std::unordered_set<std::string> names)
                    << join(names.begin(), names.end(), ", "));
 
     return result;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Get a reproducible vector of LV instance ID -> label from the given world.
- */
-std::vector<Label> make_logical_vol_labels(G4VPhysicalVolume const& world)
-{
-    std::unordered_map<std::string, std::vector<G4LogicalVolume const*>> names;
-
-    visit_volumes(
-        [&](G4LogicalVolume const& lv) {
-            // Add to name map
-            names[lv.GetName()].push_back(&lv);
-        },
-        world);
-
-    return detail::make_label_vector<G4LogicalVolume const*>(
-        std::move(names),
-        [](G4LogicalVolume const* lv) { return lv->GetInstanceID(); });
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Get a reproducible vector of PV instance ID -> label from the given world.
- */
-std::vector<Label> make_physical_vol_labels(G4VPhysicalVolume const& world)
-{
-    std::unordered_map<G4VPhysicalVolume const*, int> max_depth;
-    std::unordered_map<std::string, std::vector<G4VPhysicalVolume const*>> names;
-
-    // Visit PVs, mapping names to instances, skipping those that have already
-    // been visited at a deeper level
-    visit_volume_instances(
-        [&](G4VPhysicalVolume const& pv, int depth) {
-            auto&& [iter, inserted] = max_depth.insert({&pv, depth});
-            if (!inserted)
-            {
-                if (iter->second >= depth)
-                {
-                    // Already visited PV at this depth or more
-                    return false;
-                }
-                // Update the max depth
-                iter->second = depth;
-            }
-
-            // Add to name map
-            names[pv.GetName()].push_back(&pv);
-            // Visit daughters
-            return true;
-        },
-        world);
-
-    return detail::make_label_vector<G4VPhysicalVolume const*>(
-        std::move(names),
-        [](G4VPhysicalVolume const* pv) { return pv->GetInstanceID(); });
 }
 
 //---------------------------------------------------------------------------//
